cloud/4x_b200/requirements.txt
# 4X B200 Cloud Training Requirements
# Multi-GPU with 192GB VRAM

# Core ML libraries
torch>=2.1.0
transformers>=4.45.0
accelerate>=0.30.0
peft>=0.11.1
bitsandbytes>=0.43.0

# Data handling
datasets>=2.19.0
numpy>=1.24.0
pandas>=2.1.0

# Configuration
pyyaml>=6.0

# Web interface
gradio>=4.44.0
plotly>=5.17.0

# Hugging Face
huggingface_hub>=0.23.0
safetensors>=0.4.3

# Quantization support
bitsandbytes>=0.43.0
auto-gptq>=0.7.0

# DeepSpeed for large-scale training
deepspeed>=0.12.0
